Make base_mod iterations work correctly for the last interval

+++ src/sam/base_mods/delta.rs
@@ -1,16 +1,16 @@
 use crate::BaseModsError;
 
-pub(super) type DeltaParseFn = fn(&[u8]) -> Result<(usize, [usize; 2]), BaseModsError>;
+pub(super) type DeltaParseFn = fn(&[u8]) -> Result<(u32, [usize; 2]), BaseModsError>;
 
 pub(super) struct DeltaItr<'a> {
     data: &'a [u8],
     remaining: usize,
-    pending: Option<usize>,
+    pending: Option<u32>,
     parse: DeltaParseFn,
 }
 
 impl<'a> DeltaItr<'a> {
-    pub(super) fn new(data: &'a [u8], n_delta: usize, first_delta: usize, reverse: bool) -> Self {
+    pub(super) fn new(data: &'a [u8], n_delta: usize, first_delta: u32, reverse: bool) -> Self {
         let parse = if reverse {
             parse_mm_count_rev
         } else {
@@ -26,7 +26,7 @@ impl<'a> DeltaItr<'a> {
 }
 
 impl<'a> Iterator for DeltaItr<'a> {
-    type Item = usize;
+    type Item = u32;
 
     fn next(&mut self) -> Option<Self::Item> {
         if let Some(x) = self.pending.take() {
@@ -57,7 +57,7 @@ impl<'a> Iterator for DeltaItr<'a> {
 /// Returns error if count overflows a usize.  Return tuple with parsed count and start,stop indexes
 /// in v of remainder of input vector after removing the parsed entry (or v.len(), v.len() if there
 /// is no following entry)
-pub(super) fn parse_mm_count_fwd(v: &[u8]) -> Result<(usize, [usize; 2]), BaseModsError> {
+pub(super) fn parse_mm_count_fwd(v: &[u8]) -> Result<(u32, [usize; 2]), BaseModsError> {
     if v[0] != b',' {
         Err(BaseModsError::MissingCommaBeforeCount)
     } else {
@@ -65,14 +65,14 @@ pub(super) fn parse_mm_count_fwd(v: &[u8]) -> Result<(usize, [usize; 2]), BaseMo
             .get(1)
             .and_then(|x| if x.is_ascii_digit() { Some(x) } else { None })
             .ok_or(BaseModsError::MMCountParseError)?;
-        let mut ct = (c - b'0') as usize;
+        let mut ct = (c - b'0') as u32;
         let mut i = 2;
         if v.len() > 2 {
             for c in v[2..].iter() {
                 if c.is_ascii_digit() {
                     ct = ct
                         .checked_mul(10)
-                        .and_then(|x| x.checked_add((*c - b'0') as usize))
+                        .and_then(|x| x.checked_add((*c - b'0') as u32))
                         .ok_or(BaseModsError::MMCountParseError)?;
                 } else {
                     break;
@@ -87,16 +87,16 @@ pub(super) fn parse_mm_count_fwd(v: &[u8]) -> Result<(usize, [usize; 2]), BaseMo
 /// Parse a numeric count going backwards from the end of the slice amd ending with a comma.
 /// Returns tuple with parse count and start,stop indexes of remainder v after removing the parsed
 /// entry. Will panic if v is empty. Returns error if count overflows a usize
-fn parse_mm_count_rev(v: &[u8]) -> Result<(usize, [usize; 2]), BaseModsError> {
+fn parse_mm_count_rev(v: &[u8]) -> Result<(u32, [usize; 2]), BaseModsError> {
     assert!(!v.is_empty());
 
     let mut i = v.len() - 1;
-    let mut base: usize = 1;
+    let mut base: u32 = 1;
     let c = v[i];
     if !c.is_ascii_digit() {
         return Err(BaseModsError::MMCountParseError);
     }
-    let mut x = (c - b'0') as usize;
+    let mut x = (c - b'0') as u32;
     for c in v[..i].iter().rev() {
         i -= 1;
         if c.is_ascii_digit() {
@@ -104,7 +104,7 @@ fn parse_mm_count_rev(v: &[u8]) -> Result<(usize, [usize; 2]), BaseModsError> {
                 .checked_mul(10)
                 .ok_or(BaseModsError::MMCountParseError)?;
             x = base
-                .checked_mul((*c - b'0') as usize)
+                .checked_mul((*c - b'0') as u32)
                 .and_then(|y| x.checked_add(y))
                 .ok_or(BaseModsError::MMCountParseError)?;
         } else {
